+++ docs/assets/ai-status-indicator-BP3pFiS5.js
@@ -1,1 +1,0 @@
-import{n as e}from"./chunk-zsgVPwQN.js";import{Hm as t,Sr as n,Um as r,Wm as i,ta as a}from"./iframe-DcIOMCEi.js";var o=e((()=>{}));function s({state:e}){return(0,c.jsx)(`div`,{className:r(`jetpack-ai-status-indicator__icon-wrapper`,{[`is-${e}`]:!0}),children:(0,c.jsx)(a,{})})}var c,l=e((()=>{n(),i(),o(),c=t()}));export{l as n,s as t};